Add offline ticket creation

+++ app/jobs/reservation_cancellation_job.rb
@@ -5,10 +5,12 @@ class ReservationCancellationJob < ApplicationJob
 
   def perform(reservations)
     reservation = reservations.find(params[:id])
-    if reservation.destroy
-      render json: { success: 'Delete successful' }
-    else
-      render json: movie.errors
+    if reservation[:status] != 'paid'
+      if reservation.destroy
+        render json: { success: 'Delete successful' }
+      else
+        render json: movie.errors
+      end
     end
   end
 end
